Updated conformance tests
