+++ cabal.project
@@ -13,8 +13,8 @@ repository cardano-haskell-packages
 source-repository-package
   type: git
   location: https://github.com/input-output-hk/cardano-ledger-executable-spec.git
-  tag: ddfb7f4a311dcc52971b3bb847ba52b68d0b0a3f
-  --sha256: sha256-dQVyvzVhVUOsLUuX8XQvN/wPGmbIvYeAmX7RpnT1gk8=
+  tag: 2a8c92e0809ce703bfe8f3322085a6b9cd76e292
+  --sha256: sha256-9x8K24kgCZum+0jzVFNJTdl3BfuPb8W+EN6M3OACFUE=
 
 index-state:
   -- Bump this if you need newer packages from Hackage
