+++ eras/conway/impl/cardano-ledger-conway.cabal
@@ -115,6 +115,8 @@ library testlib
         Test.Cardano.Ledger.Conway.Binary.Cddl
         Test.Cardano.Ledger.Conway.Binary.RoundTrip
         Test.Cardano.Ledger.Conway.Binary.Regression
+        Test.Cardano.Ledger.Conway.Constrained
+        Test.Cardano.Ledger.Conway.Conformance.SpecTranslate
         Test.Cardano.Ledger.Conway.ImpTest
         Test.Cardano.Ledger.Conway.Imp
         Test.Cardano.Ledger.Conway.Imp.EpochSpec
@@ -141,14 +143,20 @@ library testlib
         bytestring,
         cardano-data:{cardano-data, testlib},
         containers,
+        constrained-generators,
+        cryptonite,
         plutus-ledger-api,
         deepseq,
         microlens,
         cardano-crypto-class,
+        cardano-crypto-wrapper,
+        cardano-ledger-executable-spec,
+        cardano-ledger-conformance,
         cardano-ledger-allegra:{cardano-ledger-allegra, testlib},
         cardano-ledger-alonzo:{cardano-ledger-alonzo, testlib},
         cardano-ledger-binary,
         cardano-ledger-babbage:{cardano-ledger-babbage, testlib},
+        cardano-ledger-byron,
         cardano-ledger-conway,
         cardano-ledger-core:{cardano-ledger-core, testlib},
         cardano-ledger-mary,
@@ -158,6 +166,8 @@ library testlib
         generic-random,
         mtl,
         text,
+        vector-map,
+        QuickCheck,
         small-steps >=1.1
 
 test-suite tests
