+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Conformance.hs
@@ -1,0 +1,6 @@
+module Test.Cardano.Ledger.Conway.Conformance (spec) where
+import Test.Cardano.Ledger.Imp.Common
+
+spec :: Spec
+spec = describe "Conway conformance tests" $ do
+  undefined
