+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Conformance/ExecutableSpecRule.hs
@@ -1,0 +1,5 @@
+module Test.Cardano.Ledger.Conway.Conformance.ExecutableSpecRule () where
+import Cardano.Ledger.Conway (ConwayEra)
+
+type instance EraFn (ConwayEra c) = ConwayFn
+
